Add files via upload
The objective of this project was to gain hands-on experience building regression models in R and to explore the statistical features of the language. Throughout the project, we created multiple regression models, checked for the normality of residuals, and assessed homoscedasticity. Additionally, we used an instrumental variable to build a regression model that accounted for endogeneity. This project allowed us to develop a deeper understanding of the various techniques and tools available in R for regression analysis, and to gain valuable practical experience in applying them.