Add files via upload
The primary focus of this project was to examine statistical regression problems in R, with a particular emphasis on utilizing prediction, logarithms, the Chow test, and confidence intervals to validate our models. Through these techniques, we gained a deeper understanding of the importance of selecting appropriate models and assessing their validity, which is crucial for generating accurate and reliable predictions. Overall, this project provided valuable hands-on experience in regression analysis in R and deepened our understanding of the fundamental concepts underlying statistical modeling.